Add app icon assets. Add scrollview and adjust layout rendering
Turns out that the UIImageView was setting the content to Aspect Fit,
and that was making the frames not line up with faces. simply setting
the content to be pinned to the top left made everything line up nice,
but then most of the image was usually not visible. Adding the
ScrollView should allow you to pan around the image, but it isn't
working yet.